init: hdp-run
