+++ cli/src/common.rs
@@ -1,12 +1,10 @@
 use alloy::{primitives::ChainId, transports::http::reqwest::Url};
 use anyhow::Result;
 use hdp::config::HdpRunConfig;
-use hdp::preprocessor::{
-    compile::config::CompilerConfig, module_registry::ModuleRegistry, PreProcessor,
-};
+use hdp::hdp_run;
+use hdp::preprocessor::module_registry::ModuleRegistry;
 use hdp::primitives::{
     aggregate_fn::{AggregationFunction, FunctionContext},
-    processed_types::cairo_format::AsCairoFormat,
     task::{
         datalake::{
             block_sampled::BlockSampledDatalake, compute::Computation, envelope::DatalakeEnvelope,
@@ -19,7 +17,6 @@ use std::{env, fs, path::PathBuf};
 use tracing_subscriber::{EnvFilter, FmtSubscriber};
 
 use clap::Parser;
-use hdp::processor::Processor;
 
 use tracing::{debug, info};
 
@@ -156,7 +153,7 @@ pub async fn module_entry_run(
     // TODO: for now, we only support one task if its a module
     let tasks = vec![TaskEnvelope::Module(module)];
 
-    handle_running_tasks(
+    hdp_run(
         config,
         tasks,
         preprocessor_output_file,
@@ -214,7 +211,7 @@ pub async fn datalake_entry_run(
         Computation::new(aggregate_fn_id, aggregate_fn_ctx),
     ))];
 
-    handle_running_tasks(
+    hdp_run(
         config,
         tasks,
         pre_processor_output,
@@ -223,67 +220,6 @@ pub async fn datalake_entry_run(
     )
     .await?;
     Ok(())
-}
-
-pub async fn handle_running_tasks(
-    config: &HdpRunConfig,
-    tasks: Vec<TaskEnvelope>,
-    pre_processor_output_file: Option<PathBuf>,
-    output_file: Option<PathBuf>,
-    cairo_pie_file: Option<PathBuf>,
-) -> Result<()> {
-    let compiler_config = CompilerConfig {
-        dry_run_program_path: config.dry_run_program_path.clone(),
-        provider_config: config.evm_provider.clone(),
-        save_fetch_keys_file: config.save_fetch_keys_file.clone(),
-    };
-    let preprocessor = PreProcessor::new_with_config(compiler_config);
-    let preprocessor_result = preprocessor.process(tasks).await?;
-
-    if pre_processor_output_file.is_none() {
-        info!("Finished pre processing the data");
-        Ok(())
-    } else {
-        let input_string = serde_json::to_string_pretty(&preprocessor_result.as_cairo_format())
-            .map_err(|e| anyhow::anyhow!("Failed to serialize preprocessor result: {}", e))?;
-        if let Some(input_file_path) = pre_processor_output_file {
-            fs::write(&input_file_path, input_string)
-                .map_err(|e| anyhow::anyhow!("Unable to write input file: {}", e))?;
-            info!(
-                "Finished pre processing the data, saved the input file in {}",
-                input_file_path.display()
-            );
-            if output_file.is_none() && cairo_pie_file.is_none() {
-                Ok(())
-            } else {
-                info!("Starting processing the data... ");
-                let output_file_path = output_file
-                    .ok_or_else(|| anyhow::anyhow!("Output file path should be specified"))?;
-                let pie_file_path = cairo_pie_file
-                    .ok_or_else(|| anyhow::anyhow!("PIE path should be specified"))?;
-                let processor = Processor::new(config.sound_run_program_path.clone());
-                let processor_result = processor
-                    .process(preprocessor_result, &pie_file_path)
-                    .await?;
-                fs::write(
-                    &output_file_path,
-                    serde_json::to_string_pretty(&processor_result).map_err(|e| {
-                        anyhow::anyhow!("Failed to serialize processor result: {}", e)
-                    })?,
-                )
-                .map_err(|e| anyhow::anyhow!("Unable to write output file: {}", e))?;
-
-                info!(
-                    "Finished processing the data, saved the output file in {} and pie file in {}",
-                    output_file_path.display(),
-                    pie_file_path.display()
-                );
-                Ok(())
-            }
-        } else {
-            Err(anyhow::anyhow!("Cairo input path should be specified"))
-        }
-    }
 }
 
 pub async fn entry_run(
@@ -326,7 +262,7 @@ pub async fn entry_run(
             }
         }
     }
-    handle_running_tasks(
+    hdp_run(
         config,
         task_envelopes,
         Some(pre_processor_output_file),
