+++ hdp/src/lib.rs
@@ -1,7 +1,10 @@
 pub mod cairo_runner;
 pub mod config;
 pub mod constant;
+pub mod hdp_run;
 pub mod preprocessor;
 pub mod primitives;
 pub mod processor;
 pub mod provider;
+
+pub use hdp_run::hdp_run;
